change route to be under traces rather than scores
