+++ client-sdks/client-js/src/resources/scores.test.ts
@@ -141,74 +141,6 @@ describe('Scores Methods', () => {
     });
   });
 
-  describe('getScoresBySpan()', () => {
-    it('should construct correct URL and make GET request', async () => {
-      mockSuccessfulResponse();
-
-      await client.getScoresBySpan({
-        traceId: 'trace-123',
-        spanId: 'span-456',
-      });
-
-      // Test that correct request is made
-      expect(global.fetch).toHaveBeenCalledWith(
-        `${clientOptions.baseUrl}/api/scores/span/trace-123/span-456`,
-        expect.objectContaining({
-          body: undefined,
-          headers: expect.objectContaining(clientOptions.headers),
-          signal: undefined,
-        }),
-      );
-    });
-
-    it('should handle HTTP errors gracefully', async () => {
-      const errorResponse = new Response('Not Found', { status: 404, statusText: 'Not Found' });
-      (global.fetch as any).mockResolvedValueOnce(errorResponse);
-
-      await expect(
-        client.getScoresBySpan({
-          traceId: 'invalid-trace',
-          spanId: 'invalid-span',
-        }),
-      ).rejects.toThrow();
-    });
-
-    it('should properly encode special characters in trace and span IDs', async () => {
-      mockSuccessfulResponse();
-
-      await client.getScoresBySpan({
-        traceId: 'trace/with/slashes',
-        spanId: 'span-with-dashes',
-      });
-
-      expect(global.fetch).toHaveBeenCalledWith(
-        `${clientOptions.baseUrl}/api/scores/span/trace%2Fwith%2Fslashes/span-with-dashes`,
-        expect.any(Object),
-      );
-    });
-
-    it('should construct correct URL with pagination parameters', async () => {
-      mockSuccessfulResponse();
-
-      await client.getScoresBySpan({
-        traceId: 'trace-456',
-        spanId: 'span-789',
-        page: 1,
-        perPage: 5,
-      });
-
-      // Test that correct URL with query params is constructed
-      expect(global.fetch).toHaveBeenCalledWith(
-        `${clientOptions.baseUrl}/api/scores/span/trace-456/span-789?page=1&perPage=5`,
-        expect.objectContaining({
-          body: undefined,
-          headers: expect.objectContaining(clientOptions.headers),
-          signal: undefined,
-        }),
-      );
-    });
-  });
-
   describe('saveScore()', () => {
     it('should save a score', async () => {
       const scoreData = {
