+++ packages/deployer/src/server/handlers/routes/observability/router.ts
@@ -1,6 +1,11 @@
 import { Hono } from 'hono';
 import { describeRoute } from 'hono-openapi';
-import { getAITraceHandler, getAITracesPaginatedHandler, processTraceScoringHandler } from './handlers';
+import {
+  getAITraceHandler,
+  getAITracesPaginatedHandler,
+  getScoresBySpan,
+  processTraceScoringHandler,
+} from './handlers';
 
 export function observabilityRouter() {
   const router = new Hono();
@@ -93,6 +98,50 @@ export function observabilityRouter() {
       },
     }),
     getAITraceHandler,
+  );
+
+  router.get(
+    '/traces/:traceId/:spanId',
+    describeRoute({
+      description: 'Get scores by trace ID and span ID',
+      tags: ['scores'],
+      parameters: [
+        {
+          name: 'traceId',
+          in: 'path',
+          required: true,
+          schema: { type: 'string' },
+          description: 'Trace ID',
+        },
+        {
+          name: 'spanId',
+          in: 'path',
+          required: true,
+          schema: { type: 'string' },
+          description: 'Span ID',
+        },
+        {
+          name: 'page',
+          in: 'query',
+          required: false,
+          schema: { type: 'number' },
+          description: 'Page number for pagination (default: 0)',
+        },
+        {
+          name: 'perPage',
+          in: 'query',
+          required: false,
+          schema: { type: 'number' },
+          description: 'Number of items per page (default: 10)',
+        },
+      ],
+      responses: {
+        200: {
+          description: 'Paginated list of scores for span',
+        },
+      },
+    }),
+    getScoresBySpan,
   );
 
   router.post(
