+++ packages/server/src/server/handlers/scores.test.ts
@@ -8,13 +8,7 @@ import { createWorkflow } from '@mastra/core/workflows';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { z } from 'zod';
 import { HTTPException } from '../http-exception';
-import {
-  getScorersHandler,
-  getScoresByRunIdHandler,
-  getScoresByEntityIdHandler,
-  getScoresBySpan,
-  saveScoreHandler,
-} from './scores';
+import { getScorersHandler, getScoresByRunIdHandler, getScoresByEntityIdHandler, saveScoreHandler } from './scores';
 
 function createPagination(args: Partial<StoragePagination>): StoragePagination {
   return {
@@ -253,138 +247,6 @@ describe('Scores Handlers', () => {
         perPage: 10,
         hasMore: false,
       });
-    });
-  });
-
-  describe('getScoresBySpan', () => {
-    it('should get scores by span successfully', async () => {
-      const mockScores = [
-        createSampleScore({ traceId: 'test-trace-1', spanId: 'test-span-1', scorerId: 'test-scorer' }),
-      ];
-      const pagination = createPagination({ page: 0, perPage: 10 });
-
-      // Mock the storage method to return our test data
-      mockStorage.getScoresBySpan = vi.fn().mockResolvedValue({
-        scores: mockScores,
-        pagination: {
-          total: 1,
-          page: 0,
-          perPage: 10,
-          hasMore: false,
-        },
-      });
-
-      const result = await getScoresBySpan({
-        mastra,
-        traceId: 'test-trace-1',
-        spanId: 'test-span-1',
-        pagination,
-      });
-
-      expect(mockStorage.getScoresBySpan).toHaveBeenCalledWith({
-        traceId: 'test-trace-1',
-        spanId: 'test-span-1',
-        pagination,
-      });
-
-      expect(result.scores).toHaveLength(1);
-
-      expect(result.pagination).toEqual({
-        total: 1,
-        page: 0,
-        perPage: 10,
-        hasMore: false,
-      });
-    });
-
-    it('should throw an error when storage method is not available', async () => {
-      const pagination = createPagination({ page: 0, perPage: 10 });
-      const mastraWithoutStorage = new Mastra({
-        logger: false,
-      });
-
-      await expect(
-        getScoresBySpan({
-          mastra: mastraWithoutStorage,
-          traceId: 'test-trace-1',
-          spanId: 'test-span-1',
-          pagination,
-        }),
-      ).rejects.toThrow(HTTPException);
-    });
-
-    it('should throw an error when storage method is not available', async () => {
-      const pagination = createPagination({ page: 0, perPage: 10 });
-      const mastraWithoutStorage = new Mastra({
-        logger: false,
-      });
-
-      await expect(
-        getScoresBySpan({
-          mastra: mastraWithoutStorage,
-          traceId: 'test-trace-1',
-          spanId: 'test-span-1',
-          pagination,
-        }),
-      ).rejects.toThrow(HTTPException);
-    });
-
-    it('should handle API errors with status codes', async () => {
-      const pagination = createPagination({ page: 0, perPage: 10 });
-      const apiError = {
-        message: 'Span not found',
-        status: 404,
-      };
-
-      mockStorage.getScoresBySpan = vi.fn().mockRejectedValue(apiError);
-
-      await expect(
-        getScoresBySpan({
-          mastra,
-          traceId: 'test-trace-1',
-          spanId: 'test-span-1',
-          pagination,
-        }),
-      ).rejects.toThrow(HTTPException);
-    });
-
-    it('should throw error when traceId is missing', async () => {
-      const pagination = createPagination({ page: 0, perPage: 10 });
-
-      await expect(
-        getScoresBySpan({
-          mastra,
-          traceId: '',
-          spanId: 'test-span-1',
-          pagination,
-        }),
-      ).rejects.toThrow(HTTPException);
-    });
-
-    it('should throw error when spanId is missing', async () => {
-      const pagination = createPagination({ page: 0, perPage: 10 });
-
-      await expect(
-        getScoresBySpan({
-          mastra,
-          traceId: 'test-trace-1',
-          spanId: '',
-          pagination,
-        }),
-      ).rejects.toThrow(HTTPException);
-    });
-
-    it('should throw error when both traceId and spanId are missing', async () => {
-      const pagination = createPagination({ page: 0, perPage: 10 });
-
-      await expect(
-        getScoresBySpan({
-          mastra,
-          traceId: '',
-          spanId: '',
-          pagination,
-        }),
-      ).rejects.toThrow(HTTPException);
     });
   });
